William work

See merge request wtraloutoure/ticp-backend!19

+++ backend/src/main/java/fr/istic/pdl/ticpbackend/service/MatchPouleService.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 
 /**
  * Ce service implémente les fonctions propres aux matchs de poule
@@ -23,7 +22,12 @@ public class MatchPouleService {
      * @return la liste des matchs de poule du tournoi
      */
     public List<MatchPoule> getMatchsPoules(){
-        return repository.findAll();
+        if(repository.findAll().isEmpty()){
+            throw new RuntimeException("Aucun match créé.");
+        }
+        else{
+            return repository.findAll();
+        }
     }
 
     /**
@@ -31,14 +35,19 @@ public class MatchPouleService {
      * @param id l'identifiant d'un match de poule
      * @return Un match de poule
      */
-    public Optional<MatchPoule> getMatchPoule(Long id){
-        return repository.findById(id);
+    public MatchPoule getMatchPoule(Long id){
+        if(!repository.existsById(id)){
+            throw new RuntimeException("Match de poule introuvable");
+        }
+        else{
+            return repository.findById(id).get();
+        }
     }
 
     /**
      * Cette méthode permet de mettre à jour un match de poule
      * @param match que l'on doit mettre à jour
-     * @throws RuntimeException si le match n'existe pas
+     * @throws RuntimeException si le match n'existe pas, si la phase de poules est terminée
      */
     public void updateMatchPoule(MatchPoule match){
         if(repository.existsById(match.getId()) & LocalDate.now().isBefore(repository.getTournoi(match.getId()).getDateFinPoule())){
@@ -52,7 +61,7 @@ public class MatchPouleService {
             throw new RuntimeException("Match inexistant");
         }
         else if(LocalDate.now().isEqual(repository.getTournoi(match.getId()).getDateFinPoule()) || LocalDate.now().isAfter(repository.getTournoi(match.getId()).getDateFinPoule())){
-            throw new RuntimeException("Phase de poule terminée");
+            throw new RuntimeException("Phase de poules terminée");
         }
 
     }
@@ -62,6 +71,11 @@ public class MatchPouleService {
      * @param id du match à supprimer
      */
     public void deleteMatchPoule(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new RuntimeException("Match de poule inexistant");
+        }
+        else {
+            repository.deleteById(id);
+        }
     }
 }
